feat: allow skipping deploy
Despite the apparent oxymoron in calling a `nomad-deploy` action and
then skipping the deployment to Nomad, this is still useful to build the
Docker container and push it to the GitHub registry, especially when in
preparation for another run of this same action (without skipping). This
flexibility is important when the same job relies on several containers
running in multiple tasks.