fix: use string-based comparison for skip-deploy
GitHub is a COOL platform with INDEPENDENT ideas, so even though the
YAML spec recognizes booleans (and editors highlight false as a
boolean), GitHub only allows for strings in inputs, meaning that the
boolean false is instead read as the string "false".

Since the string "false" is truthy, any `! skip-deploy` expressions
yield false, which is obviously contrary to our intentions. This means
one of us has to change, and since the billion-dollar company isn't, we
shall take the high road ourselves. This commit fixes this by comparing
`skip-deploy` with the string literal `true`.

Also, GitHub doesn't support "double-quoted strings" in expressions, so
we must use 'true' (vs "true") or it'll throw a syntax error. Of course.