feat!: support specifying build platforms
This allows invokers to specify a list of which platforms they wish an
image be built for, all of which are then pushed to the container
registry for downstream usage.

Although "linux/amd64" is supposedly always needed (since currently that
is what all our servers run), also building for "linux/arm64" may be
useful for some applications used in development. For example,
`bawang-content`'s Docker compose file for local development and testing
currently relies on `taitan` and `bawang`, but if those are only
published for linux/amd64, then M1/M2/M3/... Mac users will be unable to
execute it.

BREAKING CHANGE: Previously the action always built for both linux/amd64
and linux/arm64, but now the default is just linux/amd64, since the
latter platform is only needed in some repositories and thus build times
can be practically halved everywhere else by reducing the default value.